Refactor copyPackage command
- Move command to commands.ts
- Put registerCommand with the others
- Group with CFML commands in Command Palette
